adding authentication and authorization and API updates

+++ src/main/java/com/example/StocksDataAnalytics/repository/StocksRepository.java
@@ -4,5 +4,6 @@ import com.example.StocksDataAnalytics.model.Stocks;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface StocksRepository extends JpaRepository<Stocks, Long>, JpaSpecificationExecutor<Stocks> {
+
+public interface StocksRepository extends JpaRepository<Stocks, Integer>, JpaSpecificationExecutor<Stocks> {
 }
